Added a simple UI to make the use of the utility nicer.
Created a simple UI that makes the use of the utility bit nicer. Started with adding some small things to make it compatible with Linux and some initial ideas around SQL Server support. Modularized some of the PostgreSQL code further to make it easier to add support for other database engines later on.